--- API/gith/__init__.py
@@ -1,7 +1,0 @@
-'''init file'''
-
-from API.gith.actions import Actions          # noqa: F401
-from API.gith.contents import Contents        # noqa: F401
-from API.gith.issues import Issues            # noqa: F401
-from API.gith.packages import Packages        # noqa: F401
-from API.gith.pulls import PullRequests       # noqa: F401
