+++ API/github/__init__.py
@@ -1,0 +1,7 @@
+'''init file'''
+
+from API.github.actions import Actions          # noqa: F401
+from API.github.contents import Contents        # noqa: F401
+from API.github.issues import Issues            # noqa: F401
+from API.github.packages import Packages        # noqa: F401
+from API.github.pulls import PullRequests       # noqa: F401
