+++ API/github/actions.py
@@ -1,32 +1,31 @@
 '''Github API interface module'''
-# pylint: disable=R0801, R0903
-# R0801 => Calling a method shouldn't count as duplicate code
+# pylint: disable=R0903
 # R0903 => Need only 1 public method
 
 import requests
 
-from API.gith.github_base_calls import GithubBaseCalls
+from API.github.github_base_calls import GithubBaseCalls
 
 
 # Constants
-BASE_ENDPOINT = 'issues'
+BASE_ENDPOINT = 'actions'
 
 
-class Issues():
-    '''Github Issues API interface class'''
+class Actions():
+    '''Github Actions API interface class'''
 
-    def GetIssues(
+    def GetWorkflowRuns(
         owner: str,
         repo_name: str,
         is_jpl: bool = False,
         logging: bool = True
     ) -> requests.Response:
-        '''Function to call the issues endpoint'''
+        '''Function to get the workflow runs'''
 
         if logging:
             print(f'Getting data from Github "/{BASE_ENDPOINT}" endpoint...')
 
         return GithubBaseCalls.Get(
-            endpoint=f'/repos/{owner}/{repo_name}/{BASE_ENDPOINT}',
+            endpoint=f'/repos/{owner}/{repo_name}/{BASE_ENDPOINT}/runs',
             is_jpl=is_jpl,
             logging=logging)
