+++ data_updater/utils/githubutils.py
@@ -3,7 +3,7 @@
 # E1101 => dynamic method usage
 
 import json
-from API.gith import Packages
+from API.github import Packages
 
 
 class GitHubUtils():
